requirements.txt
Flask
openai-whisper
torch
numpy
